MenuServiceTest.kt 추가 및 관련 Service 동작 일부 수정.

+++ src/main/kotlin/com/hyu_oms/restapi/v5/dtos/menu/MenuListItemDto.kt
@@ -3,7 +3,7 @@ package com.hyu_oms.restapi.v5.dtos.menu
 data class MenuListItemDto(
     var id: Long,
     var name: String,
-    var price: Long,
+    var price: Int,
     var enabled: Boolean = true
 ) {
   constructor() : this(0, "", 0, true)
